+++ test/test_util.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
-from datetime import datetime, timezone
 import pytest
-from wheelodex.util import glob2like, latest_version, parse_timestamp
+from wheelodex.util import glob2like, latest_version
 from wheelodex.wheel_sort import VersionNoDot, wheel_sort_key
 
 
@@ -137,31 +136,3 @@ def test_version_no_dot(lower: str, higher: str) -> None:
 )
 def test_glob2like(glob: str, like: str) -> None:
     assert glob2like(glob) == like
-
-
-@pytest.mark.parametrize(
-    "s,dt",
-    [
-        (
-            "2018-09-26T15:12:54",
-            datetime(2018, 9, 26, 15, 12, 54, tzinfo=timezone.utc),
-        ),
-        (
-            "2018-09-26T15:12:54.123456",
-            datetime(2018, 9, 26, 15, 12, 54, 123456, tzinfo=timezone.utc),
-        ),
-        (
-            "2018-09-26T15:12:54Z",
-            datetime(2018, 9, 26, 15, 12, 54, tzinfo=timezone.utc),
-        ),
-        (
-            "2018-09-26T15:12:54.123456Z",
-            datetime(2018, 9, 26, 15, 12, 54, 123456, tzinfo=timezone.utc),
-        ),
-    ],
-)
-def test_parse_timestamp(s: str, dt: datetime) -> None:
-    parsed = parse_timestamp(s)
-    assert parsed == dt
-    assert parsed.replace(tzinfo=None) == dt.replace(tzinfo=None)
-    assert parsed.tzinfo == dt.tzinfo